opt plus works without tsp sep routine

+++ include/readers.h
@@ -10,8 +10,8 @@ using namespace std;
 
 // Function for reading the vertices files as produced by Cook, William; Hougardy, Stefan; Petrich, Moritz, 2026, "Vertices of the Subtour Polytope
 // See here https://bonndata.uni-bonn.de/dataset.xhtml?persistentId=doi:10.60507/FK2/JK95PC
-inline vector<vector<float>> read_vertices_chm(const string& filename) {
-    vector<vector<float>> vertices;
+inline vector<vector<double>> read_vertices_chm(const string& filename) {
+    vector<vector<double>> vertices;
 
     // Open the file
     ifstream vertices_file(filename.c_str());
@@ -21,7 +21,7 @@ inline vector<vector<float>> read_vertices_chm(const string& filename) {
     // Use a while loop together with the getline() function to read the file line by line
     while (getline (vertices_file, line)) {
         // Skip all the lines that begins with c
-        vector<float> vertex;
+        vector<double> vertex;
         if (not line.starts_with('c')) {
             // Split the line into tokens using space as a delimiter
             size_t pos = 0;
@@ -37,8 +37,8 @@ inline vector<vector<float>> read_vertices_chm(const string& filename) {
                 den = line.substr(2, line.find('/'));
 
                 // Convert num and den to float and push back the value num/den to the vertex vector
-                float num_float = stof(num);
-                float den_float = stof(den);
+                double num_float = stof(num);
+                double den_float = stof(den);
 
                 vertex.push_back(num_float / den_float);
             }
